# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MartHub API",
    "version": "1.0.0",
    "description": "Implemented foundation, authentication, account, and public catalog endpoints. Future inventory: docs/API.md."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/users/me/addresses": {
      "get": {
        "operationId": "addressesList",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Non-archived owned addresses, default first.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Address"
                      }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      },
      "post": {
        "operationId": "addressesCreate",
        "description": "Create an owned address. The first active address becomes default; an explicit isDefault replaces the current default atomically.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressCreate"
              },
              "example": {
                "label": "Home",
                "recipientName": "Minh Nguyen",
                "phone": "+84 912 345 678",
                "line1": "12 Market Street",
                "line2": null,
                "ward": "Ward 1",
                "district": "District 3",
                "province": "Ho Chi Minh City",
                "postalCode": "700000",
                "isDefault": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/AddressResult"
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/users/me/addresses/{addressId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressId"
        }
      ],
      "patch": {
        "operationId": "addressesUpdate",
        "description": "Update allowed fields on one owned, non-archived address. Default changes use the dedicated command.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AddressResult"
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      },
      "delete": {
        "operationId": "addressesArchive",
        "description": "Archive one owned address and remove its default flag.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Address archived. No response body."
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/users/me/addresses/{addressId}/default": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressId"
        }
      ],
      "put": {
        "operationId": "addressesMakeDefault",
        "description": "Atomically make one owned, non-archived address the sole default.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyInput"
              },
              "example": {}
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AddressResult"
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
...
    "/admin": {
      "get": {
        "operationId": "adminGetSession",
        "description": "Authorize the current database-backed Admin session and return only its safe identity projection.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current authorized Admin identity.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSessionResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
...
    "/admin/inventory": {
      "get": {
        "operationId": "adminInventoryList",
        "description": "Search and paginate single-location product inventory, optionally applying an inclusive low-stock ceiling.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "ACTIVE", "ARCHIVED"]
            }
          },
          {
            "name": "maxQuantity",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 2147483647
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 60,
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Administrative inventory list.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminInventoryListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/admin/inventory/{productId}/movements": {
      "parameters": [
        {
          "name": "productId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "adminInventoryMovementsList",
        "description": "List append-only inventory movements newest first with authoritative before and after quantities.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 60,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory movement history.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminInventoryMovementListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/admin/inventory/{productId}/adjustments": {
      "parameters": [
        {
          "name": "productId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "adminInventoryAdjust",
        "description": "Atomically apply a non-zero signed delta and append the Admin actor, reason, and quantity boundaries.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminInventoryAdjustment"
              },
              "example": {
                "adjustment": -2,
                "reason": "Damaged during cycle count"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inventory adjustment committed with its movement.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminInventoryAdjustmentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmptyInput": {
        "type": "object",
        "additionalProperties": false
      },
...
      "AddressCreate": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "recipientName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 24,
            "pattern": "^\\+?[0-9][0-9 ()-]*$"
          },
          "line1": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "line2": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 160,
                "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
              },
              {
                "type": "null"
              }
            ]
          },
          "ward": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "district": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "province": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "postalCode": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 3,
                "maxLength": 12,
                "pattern": "^[A-Za-z0-9 -]+$"
              },
              {
                "type": "null"
              }
            ]
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "label",
          "recipientName",
          "phone",
          "line1",
          "ward",
          "district",
          "province"
        ],
        "additionalProperties": false
      },
      "AddressUpdate": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "recipientName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 24,
            "pattern": "^\\+?[0-9][0-9 ()-]*$"
          },
          "line1": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "line2": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 160,
                "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
              },
              {
                "type": "null"
              }
            ]
          },
          "ward": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "district": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "province": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]+$"
          },
          "postalCode": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 3,
                "maxLength": 12,
                "pattern": "^[A-Za-z0-9 -]+$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "recipientName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "line1": {
            "type": "string"
          },
          "line2": {
            "type": ["string", "null"]
          },
          "ward": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "postalCode": {
            "type": ["string", "null"]
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "label",
          "recipientName",
          "phone",
          "line1",
          "line2",
          "ward",
          "district",
          "province",
          "postalCode",
          "isDefault"
        ],
        "additionalProperties": false
      },
...
      "AdminSessionResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "const": "ADMIN"
                  },
                  "phone": {
                    "type": ["string", "null"]
                  },
                  "status": {
                    "type": "string",
                    "enum": ["ACTIVE", "SUSPENDED", "ARCHIVED"]
                  }
                },
                "required": ["id", "email", "firstName", "lastName", "role"],
                "additionalProperties": false
              }
            },
            "required": ["user"],
            "additionalProperties": false
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
...
      "AdminInventoryListResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "sku": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": ["DRAFT", "ACTIVE", "ARCHIVED"]
                },
                "quantityOnHand": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                }
              },
              "required": [
                "productId",
                "sku",
                "name",
                "status",
                "quantityOnHand",
                "updatedAt"
              ],
              "additionalProperties": false
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "minimum": 1,
                "maximum": 9007199254740991
              },
              "perPage": {
                "type": "integer",
                "minimum": 1,
                "maximum": 60
              },
              "totalItems": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "totalPages": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": ["page", "perPage", "totalItems", "totalPages"],
            "additionalProperties": false
          }
        },
        "required": ["data", "meta"],
        "additionalProperties": false
      },
      "AdminInventoryMovementListResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "productId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "INITIAL",
                    "ADJUSTMENT",
                    "ORDER_DEBIT",
                    "ORDER_CANCEL_RESTORE"
                  ]
                },
                "adjustment": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "quantityBefore": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "quantityAfter": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "orderId": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "reason": {
                  "type": ["string", "null"]
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                },
                "actor": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        }
                      },
                      "required": ["id", "email", "firstName", "lastName"],
                      "additionalProperties": false
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "id",
                "productId",
                "type",
                "adjustment",
                "quantityBefore",
                "quantityAfter",
                "orderId",
                "reason",
                "createdAt",
                "actor"
              ],
              "additionalProperties": false
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "minimum": 1,
                "maximum": 9007199254740991
              },
              "perPage": {
                "type": "integer",
                "minimum": 1,
                "maximum": 60
              },
              "totalItems": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "totalPages": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": ["page", "perPage", "totalItems", "totalPages"],
            "additionalProperties": false
          }
        },
        "required": ["data", "meta"],
        "additionalProperties": false
      },
...
    },
    "parameters": {
...
      "AddressId": {
        "in": "path",
        "name": "addressId",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
...
      "AuthError": {
        "description": "Normalized error: INVALID_CREDENTIALS/UNAUTHORIZED (401), ORIGIN_NOT_ALLOWED (403), EMAIL_IN_USE (409), VALIDATION_ERROR (422).",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "AddressResult": {
        "description": "Owned address result without user ID or archival metadata.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "const": "no-store"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["data"],
              "properties": {
                "data": {
                  "$ref": "#/components/schemas/Address"
                }
              },
              "additionalProperties": false
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Active session and current database user status/role required."
      },
      "RefreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mh_refresh",
        "description": "Rotated opaque token. Only SHA-256 hashes are persisted; replay revokes the family."
      }
    }
  }
}